requirements.txt
altgraph==0.16.1
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.4
beautifulsoup4==4.8.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colored==1.3.93
cycler==0.10.0
Flask==1.1.1
giphy-client==1.0.0
idna==2.8
importlib-metadata==0.19
itsdangerous==1.1.0
Jinja2==2.10.1
kiwisolver==1.1.0
macholib==1.11
MarkupSafe==1.1.1
matplotlib==3.1.1
more-itertools==7.2.0
numpy==1.17.1
packaging==19.1
pandas==0.25.1
pbkdf2==1.3
pluggy==0.12.0
progress==1.5
py==1.8.0
pyaztro==0.3
pycodestyle==2.5.0
PyInstaller==3.5
pyparsing==2.4.2
pytest==5.1.1
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.2
requests==2.22.0
requests-download==0.1.2
Send2Trash==1.5.0
six==1.12.0
soupsieve==1.9.3
system==0.1.16
termcolor==1.1.0
urllib3==1.25.3
virtualenv==16.7.2
wcwidth==0.1.7
Werkzeug==0.15.6
wifi==0.3.8
zipp==0.6.0
